Allow the address to be retreived via an Endpoint.
To support a microservice scenario with a Net::Http::Endpoint, it is desirable to allow the user to specify
the port as 0, and have the operating system assign a free port.  But to do so requires reading the port
from the socket after creation, so that it can be registered with the service discovery mechanism.

The alternative of opening a socket on a random port, then reusing the port to initialise an Endpoint
exposes you to problems caused by race conditions.