Add cmake Setup Instructions For make test
make test will not work without first enabling the build of the tests.
The README.md documentation was not updated when the change that
requires the variable be enabled (863eba92) was pushed.

This resolves part of Issue #249 in that with it in place and clean
working directory the tests will compile and run.

It is probably not enough to consider #249 complete at present as all of
the test fail on my system, suggesting something is grossly wrong.