Fix Request Cookies When Receiving Multiple
Previously, when parsing the "Cookie" header for an HTTP request, the header
was parsed as if it was the "Set-Cookie" header of an HTTP response. As a
result, the CookieJar object of a Request object would never contain more than
a single cookie, with any additional cookies being stored in Cookie.ext of the
first cookie. This commit fixes this issue by implementing a
CookieJar.addFromRaw method which adds multiple Cookies to a CookieJar from a
string. HeadersStep::apply now uses CookieJar.addFromRaw instead of
Cookie::fromRaw.