Add unit tests for listener
Two unit tests were added. In the positive test, we find a port free
for usage and bind it with the listener. No exception should be
thrown. In the second test, we bind a port and then we try to bind the
same port using the listener. Here a runtime exception should be sent