Improve use of enum sets (flags)
Removes the necessity of declaring enumerators with single
bit set values (e.g. 1<<X) and the requirement of a "None"
special value (there is no such a thing as TCP option None).

Instead, operations that depend on a set of these
enumerators take a Flags<Enum> argument, which is now
implemented using a std::bitset.

Future work includes specialization of Flags such as
NotifyOn, where the value and operations make use of the
underlying C macros instead of an intermediate
bitset representation.