Creating third-party source code directory
This is a common practice int a lot of C++ projects. With this we can ignore all code analysis of third-party repos at once by removing third-party repos root path. In this directory instead of using the source itself we use a git submodule for the repository of the third-party code we want to use. In this case, I removed googletest-release-1.7.0 from root repo path and placed a git submodule pointing to official google repository in github pointing to release 1.7.0.