Creating third-party source code directory
This is a common practice int a lot of C++ projects.

The goal of this structure is to reduce complexity by ignoring third-party directory from code analysis. All repos that resides in this subdirectory would be ingored. In this case I placed googletest
inside it by linking a git submodule to google's official repository. Note that this solution could simplify the process of adding a new third party library just by creating a new submodule in this git
repository and including the code in CMakeLists.txt.