Add support for HTTP 405 responses from the REST server
If the REST router cannot find an exact match for the requested method
and resource, it will search for other methods that support the
resource. If another method supporting the resource is found, router
will send a HTTP 405 response to the client.

Also added a unit test to exercise REST server response code support.