Transport: Workaround for segmentation fault
There is a race condition where a connection being serviced can
hang up before a write operation is performed. In these cases, the
transport will issue an error response to the disconnected peer,
which invokes a promise to handle the request. The object is destroyed
and then the promise is resolved on the destroyed response object,
causing a segmentation fault