base64.cc: fix type conversion warnings
Fix type conversion warnings occurring when using -Wconversion flag
of gcc.

All the fixes actually convert std::byte to unsigned char for
calculation purposes.

Provided the number of such static_cast uses, I would reconsider using
std::byte as a base data type, as std::byte is defined as "not a
character type and is not an arithmetic type ... the only operators
defined for it are the bitwise ones". At the same time, most of the
fixes do deal with the arithmetics.

Reference:
https://en.cppreference.com/w/cpp/types/byte