Add Meson support Pistache now supports the Meson build system!
It is required to use Meson >=0.57
because previous versions can't read files,
and it wouldn't be possible to parse version.txt
to set the correct parameters in the .pc files and the shared library.
Such a recent version allows the users to
switch between the system-provided library
and the one they have in the subprojects folder
without even having to specify the fallback kwarg in dependency().

Since Meson provides builtin options for enabling code coverage
and PIC in static libraries there's no need for the PISTACHE_PIC option
and the "Profile target" mentioned by dennisjenkins75 in #840,
as they can be controlled with b_staticpic and b_coverage, respectively.

The cpp_std option is also set to c++17
even if the current minimun standard is C++14 because of #859.

If you don't know what Meson is, I highly recommend you to check it out.