Fix Debian packaging with Meson
Now the package is built with Meson. I've updated the debhelper-compat version to 12 since it provides good support for Meson, I've removed the lintian overrides (useless since the migration to Docusaurus), and obviously updated debian/rules to use Meson for building the package and creating the tarball.

This is not perfect though, as I had to use an ugly hack for the tarball since Meson does not add .github to the archive (mesonbuild/meson#8541).
Also, it is not possible to run the tests with valgrind since debhelper 12 uses ninja test for testing (while it should use meson test, so that one could add --wrap=valgrind). This behavior is corrected in debhelper 13, but it is too recent and not available in Ubuntu 20.04. It would be also possible to completely override dh_auto_test, but I don't think that's worth it.

autopkgtest does not report any error in Ubuntu Focal, so I think that everything is ok :)