ci: add conventional commits check Conventional Commits is a specification that aims at improving commit messages and history.
It consists of a simple canonical format for commit messages:
type(optional scope): description

Having a small list of types helps understand when a change is breaking,
when it adds a feature, where it patches something, or when it is
unrelated to the main library (build system changes, CI, tests...).

This is really useful when determing when to release and what type of
release should be made (major, minor, or patch release).