build(debian): use d/watch to download latest HEAD
This allows the use of uscan to get the upstream source, a much more
elegant approach.

See
manpages.debian.org/uscan.html#direct_access_to_the_git_repository_(HEAD)

I've also fixed a few Lintian warnings, like including the license with
the docs, using the wrong section for the package, not including the ITP
bug in the first changelog entry and wrong capitalization of
Vcs-Browser. The only major violation left is the use of the
"libpistache0" package name even if the SONAME is 0.0. I would suggest
removing the minor SOVERSION, as ABI is either compatible or
incompatible, there's no "minor" incompatibility. More of this in a
separate PR.