d/watch: remove manual `pretty` option
This was needed to replace the default `~git$date.$hash` suffix with
`+git$date.$hash`, because in Debian `+` is greater than `~`, and it was
needed to make the PPA happy. This unfortunately requires the d/watch
file to be updated every time the version number is bumped.

Dropping the custom version in d/watch and relying on the default is
just better, and will also fix CI.