ci: add abidiff job, remove abipkgdiff
The abipkgdiff required Debian packages to work, and while Pistache
is already buildable as a deb package with the data present in the
debian branch, building a specific commit with uscan is not really
possible.

The plain abidiff does not require packages but simple binaries, and it
works surprisingly well.

For this job I used a Debian Testing container as it is more minimal
and has access to recent packages.