ci: build on multiple Linux distributions
This new CI job builds and tests Pistache on different Linux distros,
with different compilers, and under different sanitizers, and also
re-adds Codecov coverage integration back.

Currently, Pistache is tested on Debian Stable, Debian Testing and Red
Hat Enterprise Linux 8 and 9, and the code is built with GCC and Clang.
Sanitizers are only used on Debian, because as far as I'm aware RHEL
doesn't ship the various sanitizers in its base images.

ThreadSanitizer is also disabled for now, as it reports a lot of data
races in some tests. Some of the bugs are in the unit tests themselves,
but I haven't yet looked into solving them.