test(mime): test str_to_qvalue() with bad input
As mentioned in the 5785ee1ffaa034191639e2ce223e7ec384332264 commit
message, the new str_to_qvalue() rejects malformed input. To make sure
that its behaviour won't change, I added some tests that verify that the
HttpError exeption message is the expected one, i.e.
"Invalid quality factor". This is done with GoogleTest's new
ThrowsMessage matchers. Unfortunately they are not fully documented, but
most information about them can be found in this comment:
https://github.com/google/googletest/issues/952#issuecomment-638968347

This (new?) EXPECT_THAT seems to be encouraged by GoogleTest, and I find
it quite nice:
https://google.github.io/googletest/faq.html#why-does-googletest-support-expect_eqnull-ptr-and-assert_eqnull-ptr-but-not-expect_nenull-ptr-and-assert_nenull-ptr

EXPECT_THAT is part of GMock, and the ThrowsMessage matcher is only
available in GoogleTest 1.11.0 and newer.