ci: automatically update debian/changelog
This CI job will run when version.txt is updated in the master branch,
and will push a new commit to the debian branch updating the changelog.

DEBFULLNAME and DEBEMAIL are set by "parsing" the control file. If you
know a true debian/control parser that could be used instead, preferably
part of dpkg or devscripts, please let me know.

This script currently only updates the single existing changelog entry.
It could be possible to make it add a new entry each time instead, but
it'd be a bit more involved and we would have to first check if it'd be
compatible with our current PPA setup, which I personally don't know
very well.