fix(net): drop special localhost handling
It is unnecessary to special-case the "localhost" domain name, as
documents like RFC 6761 (section 6.3)[1] already define its semantics.
It should hence be simply passed to getaddrinfo(), so that the proper
loopback address is returned.

On the other hand, using "*" as a shorthand for "0.0.0.0" is not
universally accepted, so Pistache should continue handle it specifically
for backwards compatibility.

This restores getaddrinfo() coverage regressed by commit
50e39e1afd778fed64b8925ba054e7653c58541c.

[1]: https://www.rfc-editor.org/rfc/rfc6761.html#section-6.3