refactor!: switch to sockaddr_storage
Using a struct sockaddr_storage allows us to get rid of the spooky union
of sockaddr_in and sockaddr_in6 contained in the IP class, and to do
IPv4 and IPv6 as recommended by RFC 3493.

The code is not that cleaner, but it should be more robust and less
"dangerous". As discussed recently on IRC, Kip and I were quite puzzled
by the various constructors of class IP, as it wasn't immediatly clear
why they were memcpying INET_ADDRSTRLEN bytes into a smaller s_addr
array.

Needless to say, this is a breaking change. It is impossible to change
this without breaking the ABI, and while at it I touched the API as well
to clean it up a bit.