refactor!: simplify Address::init()
In this commit I've massively refactored the init() function of the
Address class. It used to carry out most of the address parsing on its
own with a lot of custom code, making a lot of IPv4 vs IPv6 differences
and special casing. It now makes use of the library functions
standardized by RFC 3493, and most of the heavy lifting is done by
getaddrinfo().

This refactor has the following niceties:

1. Less complex code, less chances of bugs.
2. Hooking into getaddrinfo() makes Address capable of accepting
   symbolic textual service names in addition to numeric ports, so that
   users can now create an Address object like Address("*:http")
3. The loopback and wildcard address are now resolved by the system
   resolver as well, hardcoding less and yielding more desireable
   results when, for example, a system is configured to operate in IPv4
   or IPv6 only mode. Since addresses like "localhost" are now resolved
   by the system, it is possible that "::1" is now returned as well as
   "127.0.0.1", while Pistache would previously only return the latter.
   This is a change in behaviour.

The new functionalities described in points 2 and 3 are covered by new
minimal unit tests.