control: depend on libgmock-dev >= 1.11.0
Now that we have added the savoury1/build-tools PPA as a dependency of
ours, we can depend on new version of build dependencies.

This means that we can also bump other dependencies more easily, like
Meson, which would allow us to somewhat simplify our build scripts.

Note: care should be taken to not link the runtime pistache library to
dependencies only present in the build-tools PPA. It is acceptable to
use that external dependency as build-time only (e.g. tests, meson,
etc.).