fix: make zlib a private dependency
Before this patch, the zlib.h header was included by pistache/http.h,
meaning that users had to install the development headers on zlib to
build projects which depend on Pistache (in other terms, it was a
transitive, public, API dependency). Since zlib.h was included by http.h
just to get the value of zlib's default compression level, I simply
hardcoded it to 6; it's unlikely to ever change.