fix: Lock interest_mutex_ Ahead of Getting ready_ Events
We are now calling lockInterestMutex *before* we grab the contents of
ready_ in getReadyEmEventsHelper. That way, since the interest_mutex_
is required for closing an Fd, there is no risk of an Fd closing
between ready_ being grabbed (ready_ could contain the Fd_) and
lockInterestMutex being called.

Updated some debug output in os.cc.