fix: Make Sure EMEEI Valid When Close Erases From Interest
Previously, in eventmeth.cc closeEvent would erase the closing event
from interest_ and ready_ if EITHER the event was holding a pointer to
an EMEE (which has the interest_ and ready_ sets) OR global EMEE set
had an EMEE which has an interest that contained the closing
event. But this is the wrong logic. What it now does is only do the
erase if i) the event is holding the pointer to the EMEE AND ii) the
EMEE has the event in the interest set AND iii) the EMEE is in
emee_cptr_set (i.e. EMEE itself hasn't exited).

If these three conditions are not all met, closeEvent will still close
the event and delete it, but without erasing from an interest_ or
ready_ set.