fix: Use "Finalize" Flag For LibEvent Event New
Set the EV_FINALIZE flag when libevent's event_new is
called.

Otherwise libevent will have a de facto mutex on the libevent
event such that the ev_ mutex is claimed before a callback and before
allowing event_del to execute. So, when an event is being closed,
interest_mutex_ is claimed (in eventmeth), and then event_del is
called, claiming the ev_ mutex. When a callback happens, libevent
claims the ev_ mutex, and then our code claims the
interest_mutex_. I.e. the two mutexes (interest_mutex_ and ev_ mutex)
are claimed in the opposite order for a closeEvent vs. a libevent
callback, causing a deadlock on a race condition.