Move USE_LIBEVENT Detection from eventmeth.h to a New File
The code for #defining _USE_LIBEVENT, previously in eventmeth.h, was
moved to its own file emosandlibevdefs.h, which can be included
without needing to include all the rest of the eventmeth.h
capabilities.

event2/thread.h (the libevent thread header file) was also removed
from eventmeth.h. It doesn't need to be included there, and including
it there exposes eventmeth.h users directly to libevent, the opposite
of what we intend since eventmeth acts as a wrapper for libevent.