Build For FreeBSD and Fix Client Connect on Success Problem
In client.cc, the code was failing to call registerFdOneShot after
::connect returned success. Apparently in all the other *nix we've
tried, ::connect returns -1 with errno EINPROGRESS, whereas in FreeBSD
::connect simply returns 0 (success); tweaked the code so
registerFdOneShot is called in the "::connect returns 0" case. This
makes cookie_test_3 pass, it fails on FreeBSD otherwise.

emosandlibevdefs.h: Add comment on why FreeBSD uses libevent and not
the FreebSD Linuxulator.

.clang-format-ignore: Ignore the formatting in emosandlibevdefs.h
(emosandlibevdefs.h uses the same formatting as eventmeth.h - see
previous commit for why it was pulled out of eventmeth.h).

meson.build: Include execinfo library explicitly for FreeBSD (it is a
separate library in FreeBSD, but not in other *nix seemingly).

eventmeth.cc: Fixed an indentation issue (no change in functionality).