Allow UTC Suffix in Header Date String
In the test "headers_test -> last_modified_test" we allow the
last-modified-time string produced to end in either "UTC" or
"GMT". Previously, the test required "GMT". As of July/2024, it seems
that in macOS, Linux and OpenBSD it produces "GMT", while under
FreeBSD it's "UTC". Of course, they mean the same thing, and we allow
either.