Fix: In Test Code, Save Errno Before Calling strerror
Issue: In the macOS github workflow, there was an intermittent failure
for the test http_server_test /
client_request_timeout_on_delay_in_request_line_send_raises_http_408. The
test does a repeated ::send until a failure, expecting the error code
(errno saved to lastErrno_) to be EPIPE; in the failing case,
lastErrno_ is not EPIPE but zero.

Examining the code, the macro CLIENT_TRY was calling strerror to get an
error string from errno BEFORE doing lastErrno_ = errno; however this
is technically incorrect. strerror is actually allowed to modify errno
- which could result in lastErrno_ being incorrect (lastErrno_ would
hold the errno set by strerror, not the value of errno set by the
prior failing operation).

Change: in CLIENT_TRY we now do "lastErrno_ = errno" before calling strerror.

We also check for errno being zero even before strerror is called
(because the prior operation never set errno in the first place), and,
for information only, we check to see if strerror does in practice
modify errno and send a note to cout if so (if strerror does modify
errno it won't affect the subsequent code path given the change
described, but it may be interesting to know nonetheless).