fix: data race in Pistache::Rest::Router::route
This patch fixes a data race in Pistache::Rest::Router::route that
occurs when several HTTP requests are handled in parallel.
Multiple threads were writing to the same std::unordered_map
(routes) without synchronization. There is no need to modify the
map because the method is inherently constant; it’s sufficient to
find the target element in the map without adding a default value
via operator[]. Additionally, the method has been marked as const
to prevent this issue in the future.