Meson generates Windows pistache.def and pistache.lib link files
Meson Changes Explanation
=========================

In order for other programs to link with pistache.dll in Windows, they
require an import library, pistache.lib. pistache.lib defines those
DLL symbols that are accessible at link time to the DLL user.

In typical native-Windows DLL code, any symbol or class that is to be
accessible to the DLL user is decorated with "__declspec(dllexport)"
(often defined to the macro DllExport) in the source code. This
decoration is noted in the object file, and when the DLL is linked all
DllExport symbols are placed into the DLL's import library
(pistache.lib in our case) which is generated at DLL-link time.

However, Pistache assumes the Linux treatment of library symbols,
which is that any public symbol is accessible at link time to the user
of a shared library without the need to explicitly declare it for
export.

Rather than sprinkling DllExport throughout Pistache's source code, we
instead create a ".def" file which is fed to the DLL linker; the .def
file lists every public symbol of the library to be made accessible to
teh DLL user, analgously to what happens in Linux.

To achieve that, in Windows we have meson.build first build the static
version of Pistache; then we use the dumpbin utility to dump the
static library's public symbols to a file; then we use another
utility, dump2def, to derive the .def file. That .def file is then
passed to the DLL linker, which then creates the .lib file along with
creating the DLL of course.

With these changes, Pistache compiles in Windows, including the test
and example programs. A few test programs run successfully. Others
will stop with a legitimate error message when they try and access
winsock without having done WSAStartup; we haven't added a call to
WSAStartup to the code yet.

Other Changes
=============

libpistache changes:

common/http.cc:
Do integer-type casts for compression parameters, needed in Windows

Example programs:

http_server.cc
Use PST_SSIZE_T

http_server_shutdown.cc
Instead of using signal handler to indicate that a program should
close, we use a Windows-style "CtrlHandler".

Test helpers:
Link test programs with it statically, easier in Windows

Tests:

helpers_test.cc
Use em_socket_t, not int, where needed

winscripts/mestestdebug.ps1
Added to conveniently launch meson test