In Windows, Make Softlink pistache.dll for Use of Test Programs
tests/meson.build creates a softlink called "pistache.dll" that points
to the Pistache DLL in the DLL's build directory. That enables the
test program pick up the DLL from the DLL's own build directory,
without the need to install the DLL in the wider OS, which is
typically what we want when we're testing.