fix: clang++ 19 compilation error
It is not quite clear why this code does not compile only now and using
a new version of clang. Most probably, this is dead code (since using
an inaccessible/unknown `chain_` here should never compile) and clang 19
is just now a bit more aggressive in checking more things at compile
time or in a constexpr context.

If true, there is probably place to clean code further, like maybe
completely removing the `Chainer<void, Dummy>` class specialization.

But for now and without deeper knowledge or analysis this change should
not result in any behavior change and make it compile again.

Fix #1243