Feat: In Windows, Use ~/mesbuild/<projdir>/<builddir> if It Exists
The Windows convenience scripts mesbuild.ps1 and mesbuilddebug.ps1 now
check for the existence of a directory ~/mesbuild/<projdir>; if such a
directory exists, the scripts use ~/mesbuild/<projdir>/<builddir-name>
as the build direcory.

This is a way to keep the build dir far away from the sources. I use
Dropbox on the source dir, and I was finding that, in Windows, Dropbox
would occasionally interfere in the build process causing a build
error. Plus of course syncing the build dir uses unnecessary resources
and network bandwidth,